refactor: move pkg/queue to pkg/taskqueue
