+++ pkg/taskqueue/queue.go
@@ -1,4 +1,4 @@
-package queue
+package taskqueue
 
 import (
 	"encoding/json"
