+++ pkg/taskqueue/redis.go
@@ -1,4 +1,4 @@
-package queue
+package taskqueue
 
 import (
 	"context"
